+++ src/utils/shape.h
@@ -91,7 +91,7 @@ private:
   /// @note `Vector3I::dim` is used as a coordinate space dimensionality.
   static constexpr PetscInt shape_geom = POW3(shape_width);
   static constexpr PetscInt shape_comp = Vector3I::dim * 2;
-  std::array<PetscReal, (shape_geom * shape_comp)> shape;
+  std::array<PetscReal, static_cast<std::size_t>(shape_geom * shape_comp)> shape;
 };
 
 #endif  // SRC_UTILS_SHAPE_H
